data-generator/requirements.txt
snowflake-connector-python>=3.5.0
snowflake-ingest>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
faker>=20.1.0
pytz>=2023.3
requests>=2.31.0
pyyaml>=6.0
schedule>=1.2.0
py4j>=0.10.9
cryptography>=41.0.0
cffi>=1.15.0
pyarrow>=12.0.0 